feat: migrate persistence layer from sqlx to SeaORM with dual-database support
Replace the raw sqlx-based persistence layer with SeaORM, enabling
dual-database support (SQLite for dev/desktop, PostgreSQL for
production/Docker) across all 9 repository traits.

Database layer:
- Add 17 SeaORM entity definitions with typed enum columns (DeriveValueType)
  and JSON column support (FromJsonQueryResult)
- Add 14 SeaORM migrations (m000000-m000013) supporting both SQLite and
  PostgreSQL, with CITEXT/COLLATE NOCASE for case-insensitive columns
- Implement DefaultDbService as sole DbService impl wrapping
  sea_orm::DatabaseConnection with TimeService and encryption key
- Delete 28 sqlx migration files and old sqlite_pool/repository_app_instance
- Add DerivePartialModel View structs for entities with encrypted columns
  (McpAuthHeader, McpOAuthConfig, McpOAuthToken, ApiModelAlias) to prevent
  encrypted field leaks through list/get repository methods

Domain types (objs crate):
- Add DeriveValueType for domain enums (DownloadStatus, TokenStatus,
  AppStatus, McpAuthType, RegistrationType, etc.) with snake_case strum
- Add JsonVec<T> wrapper with FromJsonQueryResult for JSON array columns
- Replace uuid with ulid for ID generation (26-char Crockford Base32)
- Standardize timestamps to DateTime<Utc> with TimeService injection

Repository implementations:
- Rewrite all 9 repository impls on DefaultDbService using SeaORM queries
- McpRepository list/get methods return domain types with has_* boolean
  flags instead of Row types with encrypted fields
- Add get_decrypted_refresh_token to McpRepository trait
- ToolsetRepository create_toolset uses fetch-after-insert pattern

Service layer updates:
- Remove mcp_service row-to-model conversion helpers (repo returns
  domain types directly)
- Refactor resolve_oauth_token to use decrypt methods instead of raw
  encrypted fields
- Update access_request_service, app_instance_service, tool_service,
  queue_service for SeaORM domain types
- Update auth_middleware token_service with TimeService injection

Test infrastructure:
- Add SeaTestContext with dual-database fixture (sea_context("sqlite") /
  sea_context("postgres")) for parameterized repository tests
- Update TestDbService to wrap DefaultDbService with event broadcasting
- Add 4 new repository test files (app_instance, settings, toolset,
  user_alias)
- All repository tests run against both SQLite and PostgreSQL

E2E and CI:
- Add PostgreSQL test database to docker-compose-test-deps.yml
- Add dual-DB Playwright projects (SQLite port 51135, PostgreSQL port 41135)
- Update E2E specs for snake_case enum values and dual-DB configuration
- Update CI workflow for PostgreSQL test dependencies

Frontend and API:
- Update OpenAPI schema and regenerate TypeScript client for snake_case
  enum values
- Update UI components for RegistrationType/AppStatus enum changes

+++ crates/services/src/progress_tracking.rs
@@ -164,8 +164,8 @@ async fn update_download_progress(
     .await?
     .ok_or_else(|| format!("Download request {} not found", request_id))?;
 
-  download_request.downloaded_bytes = downloaded_bytes;
-  download_request.total_bytes = Some(total_bytes);
+  download_request.downloaded_bytes = downloaded_bytes as i64;
+  download_request.total_bytes = Some(total_bytes as i64);
   download_request.updated_at = db_service.now();
 
   if download_request.started_at.is_none() {
